requirements.txt
twilio>=9.0.0,<10.0.0
Flask>=3.0.0,<4.0.0
google-genai>=1.14.0,<2.0.0
websockets>=12.0,<16.0
python-dotenv>=1.0.0,<2.0.0
